Add Link to Join Mailing List
On the Join page, add a link to let people add themselves to the mailing
list. We would prefer of course that people fill out the full survey at
the top of the page, but the mailing list is for people who aren't ready
to fill that out but still want information on meetings, events, etc.

Resolves #17