Prevent Horizontal Scrolling on Mobile
On mobile, some text would be cut off and the screen would scroll
slightly horizontally. To fix this, we change the margins to use
`padding-left` and `padding-right`. Then, by adding
`box-sizing:border-box;`, we specify that the padding should be included
in the layout calculations. This ensures that the total width adds up to
100% of the screen width, instead of 100% being filled by the content
and then margins being added on to that.

Fixes part of #34.