Make Partner Page Mobile-Responsive
There are two levels of responsiveness here. Firstly, the `hook`, `form`,
`eligibility`, and `what-we-offer` sections are laid out like this for
desktop screens:

```
+-----------------------+
|          hook         |
+---------+-------------+
|         | eligibility |
|  form   +-------------+
|         | what-offer  |
+---------+-------------+
```

and like this on mobile:

```
+-------------+
| hook        |
+-------------+
| eligibility |
+-------------+
| what-offer  |
+-------------+
| form        |
+-------------+
```

This layout is specified by CSS Grid.

The form itself has rows laid-out using flexbot with wrapping so that
when the form is compressed too far, the input boxes drop down to
beneath the labels.

This fixes part of #34.